All lint errors fixed
* All lint errors fixed
* removed react/forbid-prop-types since PropTypes.object was used
  in 3 places
