+++ .eslintrc
@@ -1,6 +1,20 @@
 {
   "extends": "airbnb",
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "sourceType": "module",
+    "ecmaFeatures": {
+      "forOf": true,
+      "jsx": true,
+      "es6": true,
+      "experimentalObjectRestSpread" : true
+    },
+  },
   "rules": {
-    "comma-dangle": 0
-  }
+    "comma-dangle": 0,
+    "react/forbid-prop-types": [0]
+  },
+  "plugins": [
+    "react"
+  ]
 }
