+++ src/containers/App.jsx
@@ -1,30 +1,28 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import * as CounterActions from '../actions/CounterActions';
-import Counter from '../components/Counter.jsx';
-import Footer from '../components/Footer.jsx';
+import Counter from '../components/Counter';
+import Footer from '../components/Footer';
 
 /**
  * It is common practice to have a 'Root' container/component require our main App (this one).
  * Again, this is because it serves to wrap the rest of our application with the Provider
  * component to make the Redux store available to the rest of the app.
  */
-class App extends Component {
-  render() {
-    // we can use ES6's object destructuring to effectively 'unpack' our props
-    const { counter, actions } = this.props;
-    return (
-      <div className="main-app-container">
-        <div className="main-app-nav">Simple Redux Boilerplate</div>
-        {/* notice that we then pass those unpacked props into the Counter component */}
-        <Counter counter={counter} actions={actions} />
-        <Footer />
-      </div>
-    );
-  }
-}
+const App = (props) => {
+  // we can use ES6's object destructuring to effectively 'unpack' our props
+  const { counter, actions } = props;
+  return (
+    <div className="main-app-container">
+      <div className="main-app-nav">Simple Redux Boilerplate</div>
+      {/* notice that we then pass those unpacked props into the Counter component */}
+      <Counter counter={counter} actions={actions} />
+      <Footer />
+    </div>
+  );
+};
 
 App.propTypes = {
   counter: PropTypes.number.isRequired,
