+++ src/containers/Root.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { Provider } from 'react-redux';
 import App from './App';
@@ -32,6 +32,6 @@ const Root = (props) => {
 
 Root.propTypes = {
   store: PropTypes.object.isRequired,
-}
+};
 
 export default Root;
