+++ src/index.jsx
@@ -1,3 +1,4 @@
+/* global document */
 import React from 'react';
 import ReactDOM from 'react-dom';
 /**
@@ -10,8 +11,8 @@ import './styles/main.scss';
  * Both configureStore and Root are required conditionally.
  * See configureStore.js and Root.js for more details.
  */
-import { configureStore } from './store/configureStore';
-import Root from './containers/Root.jsx';
+import configureStore from './store/configureStore';
+import Root from './containers/Root';
 
 const store = configureStore();
