+++ src/store/configureStore.js
@@ -1,15 +1,43 @@
+/* eslint-disable global-require */
 /**
  * Based on the current environment variable, we need to make sure
  * to exclude any DevTools-related code from the production builds.
  * The code is envify'd - using 'DefinePlugin' in Webpack.
  */
 
-let loadedStore = null;
+import { createStore, applyMiddleware, compose } from 'redux';
+import { createLogger } from 'redux-logger';
+import thunk from 'redux-thunk';
+import DevTools from '../containers/DevTools';
+import rootReducer from '../reducers';
 
-if (process.env.NODE_ENV === 'production') {
-  loadedStore = require('./configureStore.prod');
-} else {
-  loadedStore = require('./configureStore.dev');
+/**
+ * Entirely optional, this tiny library adds some functionality to
+ * your DevTools, by logging actions/state to your console. Used in
+ * conjunction with your standard DevTools monitor gives you great
+ * flexibility!
+ */
+const logger = createLogger();
+
+const finalCreateStore = module.hot ? compose(
+  // Middleware you want to use in development:
+  applyMiddleware(logger, thunk),
+  // Required! Enable Redux DevTools with the monitors you chose
+  DevTools.instrument()
+)(createStore) : compose(
+  applyMiddleware(thunk)
+)(createStore);
+
+export default function configureStore(initialState) {
+  const store = finalCreateStore(rootReducer, initialState);
+
+  // Hot reload reducers (requires Webpack or Browserify HMR to be enabled)
+  if (module.hot) {
+    module.hot.accept('../reducers', () =>
+      store.replaceReducer(require('../reducers'))
+    );
+  }
+
+  return store;
 }
-
-export const configureStore = loadedStore;
+/* eslint-enable global-require */
