--- src/store/configureStore.prod.js
@@ -1,11 +1,0 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import thunk from 'redux-thunk';
-import rootReducer from '../reducers';
-
-const finalCreateStore = compose(
-  applyMiddleware(thunk)
-)(createStore);
-
-module.exports = function configureStore(initialState) {
-  return finalCreateStore(rootReducer, initialState);
-};
